MEJORA PARA AACEPTAR VALORES DECIMALES EN BI, GG e iva

+++ FuncionesComprobacion.c
@@ -6,7 +6,7 @@
 #include "sumatorio.h"
 
 
-int Comprobacion ( double m[], int sizeM, double DifAdm, int PorGG, int PorBI, int PorIVA, double GastosGenerales, double GastosBI, double PresupuestoPEMmasGGmasBI, double GastosIVA, double GastosPorEjecucionContrata )
+int Comprobacion ( double m[], int sizeM, double DifAdm, double PorGG, double PorBI, double PorIVA, double GastosGenerales, double GastosBI, double PresupuestoPEMmasGGmasBI, double GastosIVA, double GastosPorEjecucionContrata )
 {
 	int a = 0;
 	double sumaGG = 0;
