+++ FuncionesComprobacion.h
@@ -1,6 +1,6 @@
 #ifndef FUNCIONESCOMPROBACION_H
 #define FUNCIONESCOMPROBACION_H
-int Comprobacion ( double m[],int sizeM, double DifAdm,int PorGG,int PorBI, int PorIVA, double GastosGenerales,    double GastosBI,    double PresupuestoPEMmasGGmasBI,double GastosIVA,double GastosPorEjecucionContrata );
+int Comprobacion ( double m[],int sizeM, double DifAdm,double PorGG,double PorBI, double PorIVA, double GastosGenerales,    double GastosBI,    double PresupuestoPEMmasGGmasBI,double GastosIVA,double GastosPorEjecucionContrata );
 
 double Redondear(double x,int k);
