+++ LeerFichero.h
@@ -2,6 +2,6 @@
 #ifndef LEERFICHERO_H
 #define LEERFICHERO_H
 char* ReemplazarComaPorPunto ( char *cadena );
-void LeerficheroYSacarValores ( int argc, char **argv, double * PEMSuma, int * PorGG, int * PorBI, int * PorIVA, double  Mensualidades[],int *LongitudArray );
+void LeerficheroYSacarValores ( int argc, char **argv, double *PEMSuma, double *PorGG, double *PorBI, double *PorIVA, double  Mensualidades[],int *LongitudArray );
 #endif
